chore: general cleanup
