+++ src/util.ts
@@ -1,11 +1,6 @@
-export function stringifyError(error: unknown): string {
-  if (typeof error === "string") {
-    return error;
-  }
-  if (error instanceof Error) {
-    return error.message + "\n" + error.stack;
-  }
-  return String(error);
+export function toError(error: unknown): Error {
+  if (error instanceof Error) return error;
+  return new Error(String(error));
 }
 
 /**
